refactored into separate functions to align with ICD loader generation handle functions with no arguments
